# Speedup the Brillouin acquisition