--- a/BrillouinAcquisition/src/Acquisition/AcquisitionModes/Brillouin.h
+++ b/BrillouinAcquisition/src/Acquisition/AcquisitionModes/Brillouin.h
@@ -88,7 +88,7 @@ private:
 	int nrCalibrations = 1;
 	void calibrate(std::unique_ptr <StorageWrapper>& storage);
 
-	void abortMode() override;
+	void abortMode(std::unique_ptr <StorageWrapper>& storage) override;
 
 private slots:
 	void acquire(std::unique_ptr <StorageWrapper>& storage) override;