--- a/BrillouinAcquisition/src/Acquisition/AcquisitionModes/Fluorescence.cpp
+++ b/BrillouinAcquisition/src/Acquisition/AcquisitionModes/Fluorescence.cpp
@@ -200,7 +200,7 @@ void Fluorescence::acquire(std::unique_ptr <StorageWrapper>& storage, std::vecto
 	for (auto const& channel : channels) {
 		// Abort if requested
 		if (m_abort) {
-			this->abortMode();
+			this->abortMode(storage);
 			return;
 		}
 
@@ -270,8 +270,16 @@ void Fluorescence::acquire(std::unique_ptr <StorageWrapper>& storage, std::vecto
 	emit(s_acquisitionStatus(m_status));
 }
 
-void Fluorescence::abortMode() {
+void Fluorescence::abortMode(std::unique_ptr <StorageWrapper> & storage) {
 	m_acquisition->disableMode(ACQUISITION_MODE::FLUORESCENCE);
 	m_status = ACQUISITION_STATUS::ABORTED;
+
+	QMetaObject::invokeMethod(storage.get(), "s_finishedQueueing", Qt::AutoConnection);
+
+	// Here we wait until the storage object indicate it finished to write to the file.
+	QEventLoop loop;
+	connect(storage.get(), SIGNAL(finished()), &loop, SLOT(quit()));
+	loop.exec();
+
 	emit(s_acquisitionStatus(m_status));
 }