--- a/BrillouinAcquisition/src/Acquisition/AcquisitionModes/ODT.cpp
+++ b/BrillouinAcquisition/src/Acquisition/AcquisitionModes/ODT.cpp
@@ -195,6 +195,7 @@ void ODT::acquire(std::unique_ptr <StorageWrapper> & storage) {
 	// Here we wait until the storage object indicate it finished to write to the file.
 	QEventLoop loop;
 	connect(storage.get(), SIGNAL(finished()), &loop, SLOT(quit()));
+	QMetaObject::invokeMethod(storage.get(), "s_finishedQueueing", Qt::AutoConnection);
 	loop.exec();
 
 	m_status = ACQUISITION_STATUS::FINISHED;
@@ -327,11 +328,10 @@ void ODT::calculateVoltages(ODT_MODE mode) {
 }
 
 void ODT::abortMode(std::unique_ptr <StorageWrapper> & storage) {
-	QMetaObject::invokeMethod(storage.get(), "s_finishedQueueing", Qt::AutoConnection);
-
 	// Here we wait until the storage object indicate it finished to write to the file.
 	QEventLoop loop;
 	connect(storage.get(), SIGNAL(finished()), &loop, SLOT(quit()));
+	QMetaObject::invokeMethod(storage.get(), "s_finishedQueueing", Qt::AutoConnection);
 	loop.exec();
 
 	abortMode();