--- a/BrillouinAcquisition/src/Devices/scancontrol.h
+++ b/BrillouinAcquisition/src/Devices/scancontrol.h
@@ -133,6 +133,8 @@ protected:
 
 	void calculateHomePositionBounds();
 	void calculateCurrentPositionBounds();
+	void calculateCurrentPositionBounds(POINT3 currentPosition);
+
 
 	BOUNDS m_absoluteBounds;
 	BOUNDS m_homePositionBounds;