This PR aims to speedup the Brillouin acquisition.

Current acquisition duration:

Acquired positions and images | Nominal duration [s] | Duration with f832da60109f05b57d9699f1a5b0b6bd37712381 [s] | Duration after b78c5f3f12ca88b7b918a8a1d850325648411b20 [s] | Duration after edc48a3bc46d355aa435faa76593db70eec5183e [s]
-|-|-|-|-
0.5 s x 2 x 50 | 50 | 69 | 57 | 53
0.2 s x 2 x 50 | 20 | 38 | 27 | 23
0.1 s x 2 x 50 | 10 | 29 | 17 | 13
total overhead | - |  ~19 | ~7 | ~3
overhead per position | - |  0.4 | 0.14 | 0.06
0.5 s x 1 x 50 | 25 | 44 | 31 | 28
0.2 s x 1 x 50 | 10 | 29 | 17 | 13
0.1 s x 1 x 50 | 5 | 24 | 12 | 8
overhead | - | ~19 | ~7 | ~3
overhead per position | - |  0.4 | 0.14 | 0.06